Grammar and Name fixes
Thanks for making this website. I completely endorse it, and was really confused when I downloaded GIMP for a school project not that long ago to see it plagued with ads. Only after doing some research I found out that SourceForge is injecting adware into software that they did not wrote while proclaiming that they're simply doing the community a service by hosting mirrors.

If they wanted to host mirrors, it must be imperative that the mirrors are identical, but that's long since besides the point.


Anyway, thanks again!